[Improve]Separate the record and schema in JsonDebeziumSchemaSerializer
